pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
authors = [
    { name = "Geometric Intelligence Lab" },
    { name = "Adele Myers", email = "adele@ucsb.edu" },
]
name = "herbrain"
dynamic = ["version"]
readme = "README.rst"
description = "herbrain apps."
license = { file = "LICENSE.md" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Programming Language :: Python :: 3",
]
requires-python = ">= 3.9, <3.12"

[project.optional-dependencies]
app = [
    "typer",
    "hydra-core",
    "polpo @ git+https://github.com/luisfpereira/polpo.git@main#egg=polpo[all]",
]

[project.urls]
repository = "https://github.com/geometric-intelligence/herbrain"

[tool.setuptools.dynamic]
version = { attr = "herbrain.__version__" }

[tool.setuptools.packages.find]
include = ["herbrain", "herbrain.*"]

[tool.setuptools.package-data]
herbrain = ["*.yaml", "*.png", "*.jpeg", "*.ico"]

[tool.ruff]
extend-include = ["*.ipynb"]

[tool.ruff.lint]
extend-select = ["I", "D"]
ignore = ["E731"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
"*/__init__.py" = ["F401"]
